requirements.txt
behave
selenium
webdriver_manager
behave-html-formatter 
PyYAML
allure-behave
allure-python-commons